refactor: separate to utils
